drm/omap: Change the types of struct omap_video_timings members
omap_video_timings struct have the same members as struct videomode, but
their types are different. As first step change the types of the
omap_video_timings struct members to match their counterpart in
struct videomode to catch any type cast related issues.

+++ drivers/gpu/drm/omapdrm/displays/panel-dsi-cm.c
@@ -1024,7 +1024,7 @@ static int dsicm_memory_read(struct omap_dss_device *dssdev,
 		goto err1;
 	}
 
-	size = min(w * h * 3,
+	size = min((u32)w * h * 3,
 			dssdev->panel.timings.hactive *
 			dssdev->panel.timings.vactive * 3);
